backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pillow==10.4.0
mediapipe==0.10.18
opencv-python-headless==4.10.0.84
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
pydub==0.25.1
aiofiles==23.2.1
easydict==1.10
numpy==1.26.4
